refactor(tools): delete dead BashTool, cover #184/#220 in pty_bash
bash.rs was the hardened-but-UNREGISTERED shell tool (all_tools registers
PtyBashTool). Now that its hardening — kill-on-timeout, kill-on-drop, and
the #184 detached-grandchild non-hang — lives in the live pty_bash.rs, the
duplicate is dead code and is removed, along with its module + pub use in
lib.rs. Its #184 regression test is migrated into pty_bash.rs adapted to
the PTY tool (foreground_does_not_hang_on_detached_grandchild), and a new
timeout_kills_the_child test asserts a timed-out command's child is gone.
The #211 env-not-in-argv coverage already exists in pty_bash and is kept.
(#220)
